[FINISHES #129019501] Ajustando métodos dos controladores para serem chamados por uma classe auxiliar
