+++ grails-app/conf/laboratorios/laboratoriosDoSistema
@@ -414,4 +414,3 @@ Laboratório de Bioquímica
 Laboratório de Micologia
 Laboratório de Virologia
 NAI – Núcleo de Atendimento ao Idoso
-NewLab Test
